Hearing Added
NPCs can detect when player/other animals are within a radius of them. Doesn't actually use sound right now, just detects the object. Like with the sight, it just prints to log when something is detected.

Also did a slight change to comments on sight and fixed a line. I'm going to sleep now