Audio Source issue fixed
All the audiosources on animals and camera were set to not 3D blending and had volume distances of 500 so they could be heard at the same volume from everywhere. set spatial blend to 3D on all of them and SIGNIFICANTLY lowered the distances.

Also changed player controller so it can go up hills greater than 45 degrees